Update craftingSkill.inc
Change the skill amount when the chance is dropped significantly from 35 to 10 skill. The reason for this is that with the previous 35 setting crafting even simple things with skill requirement 0 will fail an unreasonable amount of times with ~30 skill. Another reason is that it fits well to the current setting of 10 skill maximum for vendor training. I would love to get more broad feedback on that change proposal. Are there better ways to fix this?