Added missing 'Desc' property to clothing in :clothing:itemdesc.cfg This was necessary for the guild color chooser to work properly.
Added code to :combat:equip/equip.src and :combat:equip/unequip.src to handle situations where
a guild color was changed or removed from a guild. If a guild master changes or removes
a guild's color, any clothes dyed by the guild function 'Dye clothes with the guild color.'
will either be changed to the new guild color or if the color was removed set the item's color
to zero.

Finished the guild function improvements.
These include hiding the guild functions from non-members when using the guild stone,
the ability to move a guild stone, and the ability to purchase or remove a previously
purchased guild color.

Expanded the horizontal width of the yes/no gump in :gumps:scripts/yesNo/yesNoGump.src.